+++ CotEditor/Sources/String+Counting.swift
@@ -129,3 +129,60 @@ extension String {
     }
     
 }
+
+
+
+// MARK: CharacterCountOptions
+
+struct CharacterCountOptions {
+    
+    enum CharacterUnit: String {
+        
+        case graphemeCluster
+        case unicodeScalar
+        case utf16
+    }
+    
+    
+    var unit: CharacterUnit = .graphemeCluster
+    var normalizationForm: UnicodeNormalizationForm?
+    var ignoresNewlines = false
+    var ignoresWhitespaces = false
+    var treatsConsecutiveWhitespaceAsSingle = false
+}
+
+
+extension String {
+    
+    func count(options: CharacterCountOptions) -> Int {
+        
+        guard !self.isEmpty else { return 0 }
+        
+        var string = self
+        
+        if options.ignoresNewlines {
+            string = string.replacingOccurrences(of: "\\R", with: "", options: .regularExpression)
+        }
+        if options.ignoresWhitespaces {
+            string = string.replacingOccurrences(of: "[\\t\\p{Zs}]", with: "", options: .regularExpression)
+        }
+        if options.treatsConsecutiveWhitespaceAsSingle, (!options.ignoresNewlines || !options.ignoresWhitespaces) {
+            // \s = [\t\n\f\r\p{Z}]
+            string = string.replacingOccurrences(of: "\\s{2,}", with: " ", options: .regularExpression)
+        }
+        
+        if let normalizationForm = options.normalizationForm {
+            string = string.normalize(in: normalizationForm)
+        }
+        
+        switch options.unit {
+            case .graphemeCluster:
+                return string.count
+            case .unicodeScalar:
+                return string.unicodeScalars.count
+            case .utf16:
+                return string.utf16.count
+        }
+    }
+    
+}