+++ CHANGELOG.md
@@ -7,6 +7,7 @@ Change Log
 
 ### New Features
 
+- Add Advanced Character Count to the Text menu.
 - Dynamically prioritize the scripts in the subfolder whose name is the same as the frontmost document's syntax style when the same keyboard shortcut is determined in multiple CotEditor scripts.
 - Add URL Encode/Decode commands to the Text > Transformations submenu.
 - Display a dot in the window tab if the document has unsaved changes.
@@ -19,6 +20,7 @@ Change Log
 - Update the CotEditor's setting view in the print panel.
 - Improve the algorithm to parse numbers in the Sort by Pattern command.
 - Improve the algorithm of uncommenting.
+- Deprecate the “Ignore line endings when counting characters” option.
 - [trivial] Improve the Basic Regular Expression Syntax reference.
 - [non-AppStore ver.] Update Sparkle to 2.2.0.